+++ .isort.cfg
@@ -1,0 +1,4 @@
+[settings]
+# sort and black fight over the formatting of this file.
+# Remove when this issue is resolved.
+skip=pymathics/graph/base.py
